refactor: separate DocSpaceCollapse
