+++ frontend/sprint/src/pages/TaskDetail.tsx
@@ -1,20 +1,11 @@
-import { Input, Avatar, Comment, Tooltip, Button, Collapse, Upload, message, DatePicker } from 'antd';
+import { Input, Avatar, Comment, Tooltip, Button, DatePicker } from 'antd';
 import TextArea from 'antd/lib/input/TextArea';
 import React, { useState, useEffect, createElement, useMemo } from 'react';
-import type { UploadFile, UploadProps } from 'antd/es/upload/interface';
-import { DislikeFilled, DislikeOutlined, LikeFilled, LikeOutlined, InboxOutlined, StarOutlined } from '@ant-design/icons';
-import AWS from 'aws-sdk';
+import { DislikeFilled, DislikeOutlined, LikeFilled, LikeOutlined } from '@ant-design/icons';
 import { useParams } from 'react-router-dom';
 import moment from 'moment';
 import useBindStore from '@store/zustand';
-
-interface DocumentType {
-  key: string | undefined
-  time: Date | undefined
-  url: string
-}
-
-const { Dragger } = Upload;
+import DocSpaceCollapse from '@components/DocSpaceCollapse';
 
 const TaskDetail: React.FC = () => {
   const { projectId, taskId } = useParams();
@@ -31,16 +22,6 @@ const TaskDetail: React.FC = () => {
   const [likes, setLikes] = useState(0);
   const [dislikes, setDislikes] = useState(0);
   const [action, setAction] = useState<string | null>(null);
-  const [uploadFile, setUploadFile] = useState<UploadFile[]>([]);
-  const [fileList, setFileList] = useState<DocumentType[]>([]);
-  const [uploading, setUploading] = useState(false);
-
-  const { Panel } = Collapse;
-  AWS.config.region = 'ap-northeast-2';
-  AWS.config.credentials = new AWS.CognitoIdentityCredentials({
-    IdentityPoolId: 'ap-northeast-2:89dcba84-b66a-49ce-b1f1-c0a3e77dd9da'
-  });
-  const s3 = new AWS.S3();
 
   useEffect(() => {
     setTaskInfo({ name: task?.name, content: task?.content, dueDate: task?.untilAt });
@@ -49,80 +30,6 @@ const TaskDetail: React.FC = () => {
     setEditedContent(task?.content);
     setEditedDate(task?.untilAt);
   }, [task]);
-
-  useEffect(() => {
-    s3.getSignedUrl('getObject', {
-      Bucket: 'swppsprint',
-      Key: 'asdf',
-      Expires: 604799,
-      ResponseContentDisposition: 'attachment; filename ="asdf"'
-    });
-  }, []);
-
-  useEffect(() => {
-    s3?.listObjects({ Bucket: 'swppsprint' }, (e, d) => {
-      if (e != null) void message.error('error during fetching document list');
-      const newList: DocumentType[] = [];
-      d.Contents?.forEach(file => {
-        const url = s3.getSignedUrl('getObject', {
-          Bucket: 'swppsprint',
-          Key: file.Key,
-          Expires: 604799,
-          ResponseContentDisposition: `attachment; filename ="${file.Key ?? 'asdf.txt'}"`
-        });
-        newList.unshift({ key: file.Key, time: file.LastModified, url });
-        setFileList([...newList]);
-      });
-    });
-  }, []);
-
-  const props: UploadProps = {
-    onRemove: file => {
-      const index = uploadFile.indexOf(file);
-      const newFileList = uploadFile.slice();
-      newFileList.splice(index, 1);
-      setUploadFile(newFileList);
-    },
-    beforeUpload: file => {
-      setUploadFile([...uploadFile, file]);
-      return false;
-    },
-    fileList: uploadFile
-  };
-
-  const uploadAWS = async (): Promise<void> => {
-    for (const file of uploadFile) {
-      const upload = new AWS.S3.ManagedUpload({
-        params: {
-          Bucket: 'swppsprint',
-          Key: file.name,
-          Body: uploadFile[0]
-        }
-      });
-      await upload.promise()
-        .then(() => {
-          const url = s3.getSignedUrl('getObject', {
-            Bucket: 'swppsprint',
-            Key: file.name,
-            Expires: 604799,
-            ResponseContentDisposition: `attachment; filename ="${file.name ?? 'asdf.txt'}"`
-          });
-          setFileList([{ key: file.name, time: file.lastModifiedDate, url }, ...fileList]);
-        });
-    };
-  };
-
-  const handleUpload2 = async (): Promise<void> => {
-    await uploadAWS().then(() => {
-      setUploadFile([]);
-      setUploading(false);
-    });
-  };
-
-  const handleUpload = (): void => {
-    setUploading(true);
-    void handleUpload2();
-  };
 
   const like = (): void => {
     setLikes(1);
@@ -164,8 +71,6 @@ const TaskDetail: React.FC = () => {
     setEditedName(task?.name);
     setEditedContent(task?.content);
   };
-
-  console.log(task);
 
   return (
     <div className="task-detail">
@@ -209,45 +114,7 @@ const TaskDetail: React.FC = () => {
             />
           ))}
         </div>
-        <div className="documents-container">
-          <div className="document-header">Projects Documents<Button className="document-confirm" onClick={handleUpload} disabled={uploadFile.length === 0} loading={uploading} size='small'>Confirm</Button></div>
-          {task?.document_space_list?.map(documentSpace => (
-            <Collapse accordion key={documentSpace.id}>
-              <Panel header={documentSpace.name} key={documentSpace.id}>
-                <div className="document-container">
-                  <div className="document-left">
-                    {fileList.map((file, i) => {
-                      return (
-                        <a href={file.url} className="document-uploaded" key={file.key}>
-                          <div className="file-name-container">
-                            <div className="uploaded-file">{file.key}</div>
-                            {(i === 0) && <StarOutlined size={10} />}
-                          </div>
-                          <div className="file-info">
-                            <div className="uploaded-time">{file.time?.toISOString().substring(0, 10)}</div>
-                            <div className="file-uploader">SangHyun Yi</div>
-                          </div>
-                        </a>
-                      );
-                    })}
-                  </div>
-                  <div className="document-right">
-                    <Dragger {...props}>
-                      <p className="ant-upload-drag-icon">
-                        <InboxOutlined />
-                      </p>
-                      <p className="ant-upload-text">Click or drag file to this area to upload</p>
-                      <p className="ant-upload-hint">
-                        Support for a single or bulk upload. Strictly prohibit from uploading company data or other
-                        band files
-                      </p>
-                    </Dragger>
-                  </div>
-                </div>
-              </Panel>
-            </Collapse>
-          ))}
-        </div>
+        <DocSpaceCollapse documentSpaces={task?.document_space_list ?? []} />
       </div>
     </div>
   );
